Debugged the liquibase bug and added the global exception along with the repository

+++ src/main/resources/db/changelog/changesets/03_create_table_rooms.sql
@@ -1,15 +1,15 @@
 --liquibase formatted sql
---changeset swochhandita.ghimire : rooms-v1
+--changeset swochhandita.ghimire:rooms-v1
 --preconditions onFail: HALT onError: CONTINUE
 
 CREATE TABLE rooms (
     id INT AUTO_INCREMENT PRIMARY KEY,
-    hotel_id BIGINT NOT NULL,
+    hotel_id INT NOT NULL,
     room_number VARCHAR(20) NOT NULL,
     room_type ENUM('STANDARD','DELUXE','SUITE') NOT NULL,
     status ENUM('AVAILABLE','OCCUPIED','MAINTENANCE') NOT NULL DEFAULT 'AVAILABLE',
     price_per_night DECIMAL(10,2) NOT NULL,
-    max_occupancy TINYINT NOT NULL DEFAULT 2,
+    max_occupancy INT NOT NULL DEFAULT 2,
     description TEXT,
     created_at DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
